corrections_and_create_seeders_factories

+++ database/migrations/2025_08_06_141158_create_producto_requisicion_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('id_producto')->constrained('productos')->onDelete('cascade'); #id del producto
             $table->foreignId('id_requisicion')->constrained('requisicion')->onDelete('cascade'); #id de la requisicion
-            $table->decimal('pr_amount'); #cantidad de productos por requisicion
+            $table->integer('pr_amount'); #cantidad de productos por requisicion
             $table->timestamps();
         });
     }
